v2.0.0
Removed the Balance class variable in favor of individual Checking/Savings accounts
Figured out how a Scanner issue I was having. I'm now able to have a Scanner active in the menu loop and once an option is selected in the Switch Case a new Scanner takes over. The input of which forwards to a method outside the loop with its own separate Scanner. 
This allowed me to have the menu fork from Checking/Savings to Withdraw/Deposit and they're all account specific
Balance can still be checked and now shows total for the account
Removed original deposit/withdraw methods as part of this rework.